pyproject.toml
[project]
name = "trading-system"
version = "0.1.0"
description = "AI-Powered Quantitative Trading System - 한국장 알고리즘 트레이딩"
authors = [
    {name = "Hoyoun Jung", email = "ghdbsl98@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    # 데이터 처리
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "polars>=0.20.0",  # 빠른 데이터 처리
    
    # 시각화
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.18.0",
    
    # 웹/API
    "requests>=2.31.0",
    "httpx>=0.26.0",  # async HTTP
    "aiohttp>=3.9.0",
    
    # 브라우저 자동화 (젠포트)
    "playwright>=1.40.0",

    # IPC 통신 (키움 32비트 프로세스와 통신)
    "pyzmq>=25.0.0",

    # 백테스트
    "backtrader>=1.9.78",
    "vectorbt>=0.26.0",
    
    # 데이터 소스
    "yfinance>=0.2.28",
    "pykrx>=1.0.0",  # 한국거래소 데이터
    
    # 기술적 분석
    "ta>=0.11.0",  # ta-lib 대체 (설치 쉬움)
    
    # ML (선택적 사용)
    "scikit-learn>=1.3.0",
    
    # 유틸리티
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",  # 설정/데이터 검증
    "loguru>=0.7.0",  # 로깅
    "rich>=13.7.0",  # 터미널 출력
    
    # 알림
    "python-telegram-bot>=20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "ipykernel>=6.28.0",
    "jupyter>=1.0.0",
]

ml = [
    "torch>=2.1.0",
    "tensorflow>=2.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "genport"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | backtesting-rs
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (black handles this)
]

[tool.ruff.isort]
known-first-party = ["src", "genport"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
